// src/app/model/creature.ts
export class Creature {
    "name": string;
    "level": number;
    "damageMin": number;
    "damageMax": number;
    "attack": number;
    "defense": number;
    "health": number;
    "speed": number;
    "growth": number;
    "goldPrice": number;
    "resourcePrice": string;
    "city": number;
}